rename supportedTasks.tsx to taskColumnMapping.tsx
