+++ frontend/src/components/Analysis/AnalysisTable/taskColumnMapping.tsx
